+++ src/main/resources/edu/cornell/kfs/concur/cu-spring-concur.xml
@@ -683,6 +683,12 @@
   <bean id="concurExpenseV3Service" parent="concurExpenseV3Service-parentBean" />
   <bean id="concurExpenseV3Service-parentBean" class="edu.cornell.kfs.concur.batch.service.impl.ConcurExpenseV3ServiceImpl" abstract="true">
       <property name="concurBatchUtilityService" ref="concurBatchUtilityService"/>
+      <property name="concurEventNotificationV2WebserviceService" ref="concurEventNotificationV2WebserviceService"/>
+  </bean>
+  
+  <bean id="concurEventNotificationV2WebserviceService" parent="concurEventNotificationV2WebserviceService-parentBean" />
+  <bean id="concurEventNotificationV2WebserviceService-parentBean" class="edu.cornell.kfs.concur.batch.service.impl.ConcurEventNotificationV2WebserviceServiceImpl" abstract="true">
+      <property name="concurBatchUtilityService" ref="concurBatchUtilityService"/>
   </bean>
     
 </beans>
